requirements.txt
# UI
streamlit>=1.28.0
plotly>=5.15.0    # optional, remove if no charts
schedule

# Data
pandas>=2.0.0
openpyxl
numpy>=1.24.0

# Azure SDKs & Foundry
openai
certifi
azure-ai-agents
azure-ai-projects
azure-identity>=1.14.0
python-dotenv>=1.0.0

# APIs & clients (optional but common)
fastapi
httpx
pydantic

# Testing
pytest
pytest-asyncio